pyproject.toml
[project]
name = "projeto-python-for-data-engineer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.8.5",
    "maturin[patchelf]>=1.9.4",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "polars>=1.33.1",
    "polars-hash>=0.5.5",
    "prophet>=1.1.7",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "rich>=14.1.0",
    "ruff>=0.13.0",
    "scikit-learn>=1.7.2",
    "streamlit>=1.50.0",
    "tenacity>=9.1.2",
]
